Pull empty base path test into seperate file
